Fix warnings, don't use %z in printf format (#82)
The `%z` (used for `size_t`) is not supported in C89 standard library, so avoiding it provides for better portability.

+++ core/pbcc_advanced_history.c
@@ -97,7 +97,7 @@ int pbcc_get_error_message(struct pbcc_context* p, pubnub_chamebl_t* o_msg)
                          "response='%.*s'\n",
                          p,
                          jpresult,
-                         p->http_buf_len,
+                         (int)p->http_buf_len,
                          p->http_reply);
         rslt = -1;
     }
